Code refactoring

+++ src/main/java/com/failedsaptrainees/onlinestore/services/StatisticsService.java
@@ -1,11 +1,10 @@
 package com.failedsaptrainees.onlinestore.services;
 
-import java.time.LocalDate;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.Date;
 
 public interface StatisticsService {
 
-    public Double getIncomeBetweenTwoDates(LocalDateTime fromDate, LocalDateTime toDate);
+    public BigDecimal getIncomeBetweenTwoDates(LocalDateTime fromDate, LocalDateTime toDate);
 
 }
